Move utils and clean
