+++ kvpress/__init__.py
@@ -12,7 +12,7 @@ from kvpress.presses.chunkkv_press import ChunkKVPress
 from kvpress.presses.compactor_press import CompactorPress
 from kvpress.presses.composed_press import ComposedPress
 from kvpress.presses.criticalkv_press import CriticalAdaKVPress, CriticalKVPress
-from kvpress.presses.cur import CURPress
+from kvpress.presses.cur_press import CURPress
 from kvpress.presses.decoding_press import DecodingPress
 from kvpress.presses.duo_attention_press import DuoAttentionPress
 from kvpress.presses.expected_attention_press import ExpectedAttentionPress
