+++ kvpress/presses/compactor_press.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright Vivek Chari
+# SPDX-FileCopyrightText: Copyright (c) 1993-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
 from dataclasses import dataclass
