+++ kvpress/presses/non_causal_attention_press.py
@@ -10,7 +10,7 @@ from torch.nn import functional as F
 from transformers.models.llama.modeling_llama import repeat_kv, rotate_half
 
 from kvpress.presses.scorer_press import ScorerPress
-from kvpress.presses.utils import get_query_states
+from kvpress.utils import get_query_states
 
 
 @dataclass
